breaking ParserEngine into partial files
